fix timestamp clock and read imu 6050

+++ ROS2Visualizer/app/src/main/java/com/example/ros2visualizer/model/OdomModel.kt
@@ -1,7 +1,7 @@
 package com.example.ros2visualizer.model
 
 data class OdomModel(
-    val x: Float,
-    val y: Float,
-    val theta: Float
+    val x: Double,
+    val y: Double,
+    val theta: Double
 )
